Reorganize folder structure for future changes
