+++ app/client/src/components/Admin/Shared/AdminTable.vue
@@ -1,0 +1,13 @@
+<template>
+    <div></div>
+</template>
+
+<script>
+export default {
+
+}
+</script>
+
+<style>
+
+</style>
